+++ src/app/api/popular_games/route.ts
@@ -1,33 +1,19 @@
 import { NextResponse } from "next/server";
+import { getLastThreeMonthsRange } from "./helpers/getLastThreeMonthsRange";
+
+const apiKey = process.env.RAWG_API_KEY;
+
+
+// service => /helpers/ transform... get... 
+// 
 
 export async function GET(req: Request) {
-    const apiKey = process.env.RAWG_API_KEY;
     const { searchParams } = new URL(req.url);
 
-
-    function getLastThreeMonthsRange(): { startDate: string; endDate: string } {
-        const now = new Date();
-        const endDate = new Date(now);
-        const startDate = new Date(now);
-        
-        startDate.setMonth(startDate.getMonth() - 3);
-
-        const formatDate = (date: Date): string => {
-            const year = date.getFullYear();
-            const month = String(date.getMonth() + 1).padStart(2, '0');
-            const day = String(date.getDate()).padStart(2, '0');
-            return `${year}-${month}-${day}`;
-        };
-
-        return {
-            startDate: formatDate(startDate),
-            endDate: formatDate(endDate),
-        };
-    }
-
     const { startDate, endDate } = getLastThreeMonthsRange();
-    
+
     const dates = `${startDate},${endDate}`
+
     const page_size = searchParams.get('page_size') || '8';
 
     try {
